+++ import_export/admin.py
@@ -753,7 +753,7 @@ class ExportMixin(BaseExportMixin, ImportExportMixinBase):
             form.fields["export_items"] = MultipleChoiceField(
                 widget=MultipleHiddenInput,
                 required=False,
-                choices=[(pk, pk) for pk in self.get_valid_export_item_pks(request)],
+                choices=[(pk, pk) for pk in queryset.values_list("pk", flat=True)],
             )
         if form.is_valid():
             file_format = formats[int(form.cleaned_data["format"])]()
@@ -774,30 +774,6 @@ class ExportMixin(BaseExportMixin, ImportExportMixinBase):
         context = self.init_request_context_data(request, form)
         request.current_app = self.admin_site.name
         return TemplateResponse(request, [self.export_template_name], context=context)
-
-    def get_valid_export_item_pks(self, request):
-        """
-        DEPRECATED: This method is deprecated and will be removed in the future.
-        Overwrite get_queryset() or get_export_queryset() instead.
-
-        Returns a list of valid pks for export.
-        This is used to validate which objects can be exported when exports are
-        triggered from the Admin UI 'action' dropdown.
-        This can be overridden to filter returned pks for performance and/or security
-        reasons.
-
-        :param request: The request object.
-        :returns: a list of valid pks (by default is all pks in table).
-        """
-        cls = self.__class__
-        warnings.warn(
-            "The 'get_valid_export_item_pks()' method in "
-            f"{cls.__module__}.{cls.__qualname__} "
-            "is deprecated and will "
-            "be removed in a future release",
-            DeprecationWarning,
-        )
-        return self.model.objects.all().values_list("pk", flat=True)
 
     def changelist_view(self, request, extra_context=None):
         if extra_context is None: